requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-filter==23.1
requests==2.26.0
psycopg2-binary>=2.9
flake8>=4.0.1
ruff>=0.1.7
pytest>=7.0.0
pytest-django>=4.5.2
gunicorn>=20.1.0
drf-yasg>=1.21.5